requirements.txt
-i https://mirrors.aliyun.com/pypi/simple
aiohttp==3.8.1
aioredis==2.0.1

bs4==0.0.1

celery==4.3.0

elasticsearch-dsl==7.4.0

flask==1.1.2
flake8==3.9.0
fastapi==0.73.0

gevent==21.1.2
greenlet==1.0.0
glom==22.1.0

httpx==0.22.0

lxml==4.6.3

matplotlib

numpy==1.19.5

#pip==22.0.4
pyquery==1.4.3
pandas==1.1.5
pydantic==1.9.0
peewee==3.14.8
pymysql==1.0.2
pymongo==4.0.1
pillow==8.2.0
#pycrypto

requests==2.25.1
redis==4.1.2
requests_toolbelt==0.9.1
python-redis-lock[flask]==3.7.0

scipy==1.5.4
selenium==3.14.1

tqdm==4.59.0

vine==1.3.0


















